server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Fridge2Fork Server - Recipe Recommendation API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-jazzmin>=3.0.1",
    "django-ninja>=1.4.3",
    "dotenv>=0.9.9",
    "gunicorn>=21.2.0",
    "ipykernel>=6.30.1",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
]
